UT: Move a ransomware test to integration_tests
