+++ monkey/tests/unit_tests/infection_monkey/payload/ransomware/conftest.py
@@ -11,11 +11,6 @@ def patched_home_env(monkeypatch, tmp_path):
     return tmp_path
 
 
-@pytest.fixture(params=[".m0nk3y", ".test", ""], ids=["monkeyext", "testext", "noext"])
-def ransomware_file_extension(request):
-    return request.param
-
-
 @pytest.fixture
 def ransomware_test_data(data_for_tests_dir):
     return Path(data_for_tests_dir) / "ransomware_targets"
